+++ packages/react-changelogs/examples/filter-changelogs.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import styled from 'styled-components';
+import styled from '@emotion/styled';
 import { gridSize, math } from '@atlaskit/theme';
 import Changelog from '../src/components/changelog';