+++ packages/website/src/index.ts
@@ -1,0 +1,3 @@
+import runPipeline from './pipeline';
+
+export default runPipeline;